STY: Moved input coordinate test function
